refator events: move temperature calibration implementation into subdirectory
