+++ src/modules/events/CMakeLists.txt
@@ -38,9 +38,9 @@ px4_add_module(
 	COMPILE_FLAGS
 	SRCS
 		send_event.cpp
-		temperature_gyro_calibration.cpp
-		temperature_accel_calibration.cpp
-		temperature_baro_calibration.cpp
+		temperature_calibration/gyro.cpp
+		temperature_calibration/accel.cpp
+		temperature_calibration/baro.cpp
 	DEPENDS
 		platforms__common
 		modules__uORB
